Step 1: ASTs and parsing
Before we define our Joy quotation we need to implement syntax trees and
a parser for the language.

The subset that we'll be working with contains: integer literals,
boolean literals, names (which represent calls to predefined
operators) and quotations. Plus there is also an extra Program
node that represents whole code snippet.

In an imaginary Joy repl those could be used in the following way:

    > 1 1 +
    2

    > 1 2 =
    false

    > true false or
    true

    > 2 2 2 * *
    8

    > [1 2 3 + +] i
    6

    > [1 2 3] first
    1

    > [1 2 3] rest
    [2 3]

    > 2 [0 >] [true] [false] ifte
    true

    > [1 2 3] [1 +] map
    [2 3 4]

    > [1 2 3 4] [2 rem 0 =] filter
    [2 4]

Parser is implemented via Scala's parser combinators but just about any
parsing library would do. All you really need is a String => Joy mapping.