Step 4: moving parsing to compile-time
We use macros to remove all of the runtime parsing overhead. Apart from
moving code to the macro implementation we also need to implement
Joy tree to Scala tree conversion (called lifting). This is needed
because macro has to return a tree that represents a value rather
than a value itself.

It's important to highlight that we opt in to use whitebox macros rather
than blackbox ones as it will give as ability to refine return type of
the quotation to specific subtypes of Joy tree.

For example in a few steps we'll make following work:

    scala> j"42"
    res0: joy.Joy.Int = 42

Here res0 is Joy.Int rather than just Joy thanks to whiteboxity.