Step 6: unquoting
The problem with our current quotation is that it doesn't provide any
means to compose bigger ASTs from smaller chunks.

The solution is unquoting. Whenever user writes $expr in a quotation it
should be interpreted as a structural substitution of expr value in that
position.

To implement this we need to parse all string parts rather than just the
first one. To fill links between them we insert special mangled names
(which are constructed and reconstructed via Hole object). These names
will let us correlate location of trees that correspond to $unquotes and
replace them with corresponding argument during lifting.