Step 11: basic pattern matching
Our quotation so far only worked in expression mode and failed to
compile if it was used as a pattern.

Now we augment the definition of the interpolator with pattern
matching-friendly unapply entry point which allows our macro to expand
in pattern position.

Apart from second entry point we also need to include a different
wrapping strategy in pattern position which is caused by the interaction
between macros and pattern matching.

Interestingly most of the lifting code can stay the same due to the fact
that untyped trees for case class creation and pattern matching are the
same but there is a change we need to make to make list lifting
work.

Previous calls to :: and :+ methods needs to be replaced with
explicit call to :: list constructor and a utility joy.` :+ ` which works
similarly to :: but only for appends. We can't use default :+ object
because it doesn't have .apply method and won't work in both contexts.