Step 12: pattern unquoting and splicing
Now that we've got simple patterns working we need to take care of
unquoting and splicing.

First implmentation of arg(i) in expression mode could just returns args(i)
but the situation is different here. Due to the fact that macros in
pattern mode don't rewrite original pattern but just augment it with
generated extractor we need to generate a bit more code.

To implement this we create a new indexed variable binding for each argument in
arg(i) and than reconstruct a tuple of all values on the right hand
side wrapped into Some. For example following pattern:

    case j"$a $b +" =>

Will have following generated extactor:

    new {
      def unapply(input: Joy) = input match {
        case Joy.Program((x1 @ _) :: (x2 @ _) :: Joy.Name("+") :: Nil) =>
          Some((x1, x2)
        case _ =>
          None
      }
    }.unapply(...)

If there is no arguments then right hand side of the generated
cases will contain true for the first case and false for the other one.

Luckily due to usage of :: constructor and ` :+ ` extractor we
automatically get support for splicing without any additional
changes.